Report on get_childrent function call due to default -1 for posts_per_page
Since the `get_children` is `setting posts_per_page` to `-1` by default, we should make sure that developers are overriding the default value to something reasonable - eg.: 100

This commit is changing the parent of the `WordPressVIPMinimum_Sniffs_VIP_RestrictedFunctionsSniff` class from `WordPress_AbstractFunctionRestrictionsSniff` to `WordPress_Sniffs_VIP_RestrictedFunctionsSniff` and takes advantage of the inheritance which should allows us to override the original mesages in future.

Further, it is adding the `get_children` group with a message about the default `-1` value for `posts_per_page`.

Fixes #10