PHPCS 3.x compatibility changes
This commit makes changes necessary for succesffully runnig the WordPressVIPMinimum standard sniffs against both, PHPCS 2.x and 3.x.

Due to the project's dependency on the WPCS, it is now loading the backward compatibility code from the upstream repository and enhances it for some missing parts from there.

All tests are now also aliasing the `AbstractSniffUnitTest` if needed - when run against PHPCS 3.x.