Add minimum version requirements of PHPCS and WPCS
Despite my best efforts, I could not make the VIP Coding Standards working on arbitrary WPCS and PHPCS version - however, they do work on the latest and possibly the last version of the 2.x branch of the PHPCS (v2.9.1). When it comes to WPCS, the 0.13.1 (currently the latest one), is the minimal required version.