Warn, rather than error, for deprecated VIP helpers.
The presence of VIP helpers aren't strictly blockers because they won't cause any serious performance issues nor are they a security risk. It is important that these are replaced with the core alternatives, but not critical.

This change bumps these down to a warning, and more inline with what PHP or WordPress would report when a deprecated function is used.