PHPCS: Add XSD reference and validate rulesets
As of PHPCS 3.2.0, PHPCS includes an XSD schema for rulesets which defines what can be used in the XML ruleset file.

In PHPCS 3.3.0 a new array format for properties was introduced and as of PHPCS 3.3.2, the new array format is now also covered by this schema.

This commit:

- Adds the schema tags to the `ruleset.xml` files, and the VIPCS native `.phpcs.xml.dist` file.
- Adds validation against the schema for the `ruleset.xml` files to the Travis build.
- Fixes invalid markup in the `WordPress-VIP-Go` ruleset flagged by the validation procedure.

Note:
The schema used in the tags points to the online schema in PHPCS master for two reasons:

1. While VIPCS has a minimum requirement of PHPCS 3.2.3, the schema in PHPCS master contains the changes which were made in PHPCS 3.3.2, so we can safely validate the rulesets against the schema.
2. While installation via Composer is supported for WPCS, it is not the only supported installation method, so pointing to the version of the XSD schema in the vendor directory would be presumptuous.

At the same time, using the URL in the Travis script appears to be problematic so we do a `composer require` for PHPCS for the actual validation. As the validation is done on a build with PHPCS master, there will be no difference anyway.

Fixes #302.

Much kudos to @jrfnl for her work and words that this commit is highly based on.