RestrictedFunctions: Clarify object variable usage
Make it clear that the method call is actually on a specific _object variable_, regardless of what class that object is an instantiation of. Indicate that this works best for global object variables, since any change of _variable name_ (reassigning global or other object to a different variable) will cause this check to miss.

See #279.