Remove CronInterval in favour of custom property
The VIPCS CronIntervalSniff class was added to re-assert the minimum cron interval is 900 seconds.

This already is the default minimum interval, but since it is a public property in the WPCS CronIntervalSniff class, we can re-assert it via a custom property in the ruleset, rather than adding a new sniff class.

- Removes sniff class file
- Removes sniff test files
- Adds test code into integration test file and amends integration test line numbers.
- Adds `min_interval` property to `WordPressVIPMinimum` ruleset, in case the default ever changes in WPCS.

Fixes #324.