Copy AbstractVariableRestrictionsSniff from WPCS
WPCS 1.0 marked the `AbstractVariableRestrictionsSniff` as deprecated, since the only sniffs using it were from the `WordPress.VIP` category of sniffs, which were themselves deprecated.

WPCS 2.0 removes the `AbstractVariableRestrictionsSniff`, so this copies it into VIPCS where we can maintain and improve it.

It is currently only extended by the RestrictedVariablesSniff class.

Fixes #325.