Refactor sniffs to extend from Sniff
One initial benefit of the sniffs extending from a class that extends from `\WordPress\Sniff`, is that the sniffs can call `process_token()`, which is what is called in \WordPress\Sniff::process() [1] after the `$phpcsFile` and `$tokens` properties have been populated. This reduces the amount of boilerplate code in each sniff, and ensures there is a consistent presence and naming of sniff properties that contain references to the files and the tokens stack.

[1]: https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/blob/b6d70528648a59434bfbf379dcae876f6c9abe32/WordPress/Sniff.php#L817

Fixes #359.