Add WordPressVIPMinimum\Sniffs\Sniff
This serves as a base class from which all other VIPCS sniffs should
extend. It extends `\WordPress\Sniff` (and will be the only class that
does, to keep the coupling in a single place) so that it, and all of the
VIPCS classes can use the utility methods that `\WordPress\Sniff`
contains. This in turn helps us to improve the quality of our sniffs,
and reduce the amount of logic there is to maintain.

Although empty now, we can introduce our own utility methods to this
class.

See #359.