HTMLExecutingFunctions: Add more functions
There are different ways to insert variables into the DOM, and our sniffs only covered some of them.

The sniff now has an expanded list, which covers functions that have a syntax where the content is in the function arg, and also where the function arg is the target and the content is in the preceding method's arg.

Fixes #435.