append: Remove error for use of append in VIP Go
It's not clear why this violation is marked as an Error for VIP Go ruleset, when none of the other HTMLExecutingFunctions are left as their default of Warnings.

Since the best the Sniff can do is provide warnings when a variable is used (though that variable may already be safe, and not contain user input), then it would seem odd to change just one function call to an Error for VIP Go.