Redirect functions: check for third parameter
Debugging redirects can be hard as there are so many levels where a redirect may happen.

WordPress 5.1 and later adds an `X-Redirect-By` header when `wp_redirect()` or `wp_safe_redirect()` is used, with a default value of `WordPress`, but this can be customised to any string value by adding a third argument to those function calls. This would make debugging redirects easier.

This PR adds a Warning if the third arg is missing.

Fixes #433.