Contributing: improve the text
As GH was so kind as to point me to the "contributing guidelines which have been changed since you last contributed" I quickly scanned through the file and most notably, came across a link to the WPCS wiki page, while this link should have gone to the VIPCS wiki.

As I was editing the file now anyway, I've made some other changes along the way.

* Ask people to search the issue list before opening an issue.
* Made a little table to make it easier for people to report upstream issues to the correct repository.
* Updated the "custom properties" link to point to the VIPCS wiki.
* Updated the PHPUnit installation link.
* Updated the `phpunit.xml` example to be in line with the current `phpunit.xml.dist` file.
* Updated the expected unit test output.