WordPressVIPMinimum: fix the failing ruleset test
Ok, this is a fun one.

Line 5 of the ruleset-test is not supposed to error. It is just setting up the variables for the VariableAnalysis test.

Line 7, however, is supposed to error on a PHP parse error (`Generic.PHP.Syntax`).

The above is true and working for PHP 7.

However on PHP 5, line 7 will not throw a parse error, while line 5 actually will - `Cannot re-assign $this`.

To fix this, I've:
* Replaced the `$this` variable assignment on line 5 with an assignment to `$stdClass`.
* Replaced the PHP 7 parse error on line 7 with a parse error which will error on all currently known PHP versions.

Fun times ;-)