DeclarationCompatibility: fix incorrect signature check for `Walker::walk()`
... and `Walker::paged_walk()`.

The method signature of both the `Walker::walk()` method as well as the `Walker::paged_walk()` method were changed in WordPress 5.3 to make the variadic nature of the allowed arguments explicit.

This changes update the sniff to verify the signature of these methods in child classes against the new method signature as it is in WordPress Core.

Refs:
* https://make.wordpress.org/core/2019/10/09/wp-5-3-introducing-the-spread-operator/
* https://core.trac.wordpress.org/changeset/46442

Fixes 448