RulesetTest: don't use the system default version of PHP
... instead use the PHP version used when this script was called.

`shell_exec()` opens a new process and that process does not inherit the PHP executable which was used to run the tests, but uses the system default PHP version instead, which means that the PHPCS command may run on a different PHP version than the tests.

The current solution is a simple one and not exhaustive, but should suffice in most cases.

If a more complex/more comprehensive solution is desired, I'd recommend using the `PhpExecutableFinder` from the `Symfony\Process` package.