PreGetPosts: improve the isEarlyMainQueryCheck() method [2]
This adds a second unit test which is based on the actual code which originally triggered the error.

The error was caused by the code in question using inline control structures (without braces) for the early main query check.

After the previous fix, that code would now throw a false positive.

I've fixed this now by adding an additional check for a `return` statement straight after the parenthesis closer of the `if()` statement.

Fixes 499