RestrictedFunctions: fix false negative - functions in `config_settings` would never match
The functions in the `config_settings` group were using a unicode `_`, not the ascii `_`, so these would never match the PHP native `opcache_*` functions they were intended to match.

As the same unicode character was used in the applicable unit tests and ruleset tests, the bug was never reported as a test failure.