AlwaysReturnInFilter: differentiate between abstract method and parse error
This implements the suggestion made in https://github.com/Automattic/VIP-Coding-Standards/issues/580#issuecomment-689119006

If the method a filter callback points to is an abstract method, a warning will be thrown asking for manual inspection of the child class implementations of the abstract method.

In case of parse or tokenizer error, the sniff will bow out and stay silent.