Functions/DynamicCalls: bug fix - don't blindly use the next text string
A variable value may be build up of multiple tokens.

As it was, the sniff would look for the first text string token after the equal sign within the variable assignment statement, but this disregards that:
1. The text string token found may not be the only token in the statement.
2. A statement can end on a PHP close tag (possibly a bug in PHPCS itself, but that's another matter), which would lead the sniff to look at the next statement for text strings.

Fixed now.

Includes unit tests, the first four of which resulted in false positives previously.