Functions/DynamicCalls: bug fix - fix memory + performance issue
The sniff maintains a cache of all the variables it has seen and their assigned value.

When a variable is encountered, it would:
* Check if it was an (plain text) assignment and if so, register the variable name + value to the cache.
* Next, call the `find_dynamic_calls()` method, which first checks if any variables have been registered to the cache before doing anything.
* And then checks for dynamic function calls and if one is found, checks if the variable used is one registered in the cache with a value we are looking for.

This is highly inefficient as text string variable assignments are common and, as it was, _every single one_ would be added to the cache.

With a large code base, that means that the cache could grow pretty large.

It also means that the logic to determine if something is a dynamic function call would be executed even when there would be no text strings registered in the cache which could match any of the ones we're looking for.

By changing the order of the logic, the memory leak and performance inefficiency is removed.

With the updated logic, the sniff will:
* Check if it was an (plain text) assignment **and if the text string matches one we're looking for** and if so, register the variable name + value to the cache.
* Next, call the `find_dynamic_calls()` method, which first checks if any variables have been registered to the cache before doing anything.
* And then checks for dynamic function calls.

This means that if none of the previous assignments encountered matches any of the target text strings (~ 99% of the time), this sniff will bow out at step 2 before executing the logic to check if a variable assignment is a dynamic function call.