Travis: add build against PHP 8.0
PHP 8.0 has been branched off two months ago, so `nightly` is now PHP 8.1 and in the mean time PHP 8.0 was released last week.

As of today, there is a PHP 8.0 image available on Travis.

This PR adds a two new builds against PHP 8.0 to the matrix and, as PHP 8.0 has been released, these builds are not allowed to fail.

PHPCS 3.5.7 is the first PHPCS version which is runtime compatible with PHP 8.0.
Once the minimum supported PHPCS version of this package has been upped to PHPCS 3.5.7, these builds don't need to be special cased anymore.