ProperEscapingFunction: only trigger on function calls
While unlikely, a class or constant could be declared which uses the same name as one of the target functions.
In that case, the sniff should not trigger an error.

This adds a safeguard to verify if something is a function call and bows out if not.

Includes unit test.