ProperEscapingFunction: move a condition
As the other checks now each have a condition checking the `$escaping_type` in the `process()` method, having that condition in the `is_outside_html_attr_context()` function for the "esc_attr being used outside HTML tags" check makes it inconsistent and raises the cognitive complexity.

It is more intuitive to have a similar "prelim function type" check in the condition in the `process()` method.

This also removes the need to pass the `$function_name` to the `is_outside_html_attr_context()` function.